% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_key.R
\name{annotate_key}
\alias{annotate_key}
\title{annotate_key}
\usage{
annotate_key(key)
}
\arguments{
\item{key}{a key created by Vizumap::build_bkey}
}
\value{
the key with annotations; of class "bivkey"
}
\description{
Adds key annotations that correspond to the bin's position on the x-axis
}
\examples{
out = st_assign_colors(tmax, 1, tmin, 1, terciles = TRUE)
data = out$data
colors = out$colors

key = Vizumap::build_bkey(data, colors, terciles = TRUE)
annotate_key(key)
}
