% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalize.R
\name{anomalize}
\alias{anomalize}
\title{Anomalize Datasets}
\usage{
anomalize(starsObj, attr, space1, space2 = NULL, time, baseline = NULL, ...)
}
\arguments{
\item{starsObj}{a stars object}

\item{attr}{attribute of stars object}

\item{space1}{spatial dimension (e.g., station, Longitude)}

\item{space2}{second spatial dimension, if applicable (e.g., Latitude)}

\item{time}{time dimension}

\item{baseline}{optional parameter that chooses which time period to calculate anomalies for (e.g., 1951-1980)}

\item{...}{other dimensions not involved in anomalization, but still part of the original stars object}
}
\value{
a stars object with anomalized attribute
}
\description{
Redefine data points as how far away they are from the average for that point in space, a useful tool for better visualization and analysis
}
\examples{

}
