% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_colors_df.R
\name{assign_colors.df}
\alias{assign_colors.df}
\title{Bin bivariate ST data and assign colors}
\usage{
assign_colors.df(data, bounds, palette = NULL)
}
\arguments{
\item{data}{a data frame with the variables of interest in the first two columns.}

\item{bounds}{How should the data be binned? Use custom bounds or those returned from st_findNbounds.}

\item{palette}{A palette built using Vizumap::build_palette or st_build_palette.}
}
\value{
data with two extra columns: bothVars, a factor, and hex_code, the color corresponding to each
data point.
}
\description{
Bin bivariate ST data and assign colors
}
\examples{
temps.DF = as.data.frame(tmax)
names(temps.DF)[3] = "tmax"
tmin.DF = as.data.frame(tmin)

temps.DF$tmin = tmin.DF$tmin.nc
temps.DF = temps.DF[,c(3,4,1,2)]

bounds = st_findNbounds(temps.DF)
temps.DF = assign_colors.df()

}
