% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_key.R
\name{attach_key}
\alias{attach_key}
\title{View bivariate histograms with their keys}
\usage{
attach_key(plot, VizuKey, draw = TRUE)
}
\arguments{
\item{plot}{a plot such as those returned from st_bivariate_histogram()}

\item{VizuKey}{a key built with Vizumap::build_bkey()}

\item{draw}{should the plot be drawn on the current device (default)? if FALSE, returns a grob without drawing. See gridExtra::arrangeGrob()}
}
\value{
draws on the current device or returns a grob
}
\description{
View bivariate histograms with their keys
}
\examples{
# Preparing the data
out <- st_assign_colors(tmax, 1, tmin, 1, terciles = FALSE)
data <- out$data
colors <- out$colors
bounds <- out$bounds

# generating and plotting the histogram
bvh <- st_bivariate_histogram(data = data,
            title = "Max/Min temperature distribution")
bvh

# creating and attaching the key
key <- Vizumap::build_bkey(data, colors, terciles = TRUE)
Vizumap::view(key)
attach_key(bvh, key)
}
