% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcbin.r
\name{calcbin}
\alias{calcbin}
\title{calcbin}
\usage{
calcbin(terciles, data, x, q, bin, width, min)
}
\description{
Code extracted from Vizumap to compute the bins from which
the 3x3 grid is produced.
}
\details{
Permission granted from authors under GPL
}
