% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNbounds_helper.r
\name{findNbounds_helper}
\alias{findNbounds_helper}
\title{findNbounds_helper}
\usage{
findNbounds_helper(data, estimate, error, terciles)
}
\description{
Code extracted from Vizumap to find the bounds of the data.
Permission granted from authors under GPL
}
