% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_density.r
\name{get_density}
\alias{get_density}
\title{Get 2d density}
\usage{
get_density(x, y, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}

\item{...}{Create a square n by n grid to compute density.}
}
\value{
The density within each square
}
\description{
A convenience function to find 2d densities using MASS::kde2d.
get_density function found at https://slowkow.com/notes/ggplot2-color-by-density/
}
