% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GISTEMP}
\alias{GISTEMP}
\title{Global Temperature Anomaly Data.}
\format{
A stars object with 3 dimensions and 1 attribute:
\describe{
\item{Temperature Anomalies}{How far a measurement departs from 1920-1950 mean, in degC.}
\item{Long}{Longitude, in degrees}
\item{Latitude}{Latitude, in degrees}
\item{Time}{Year}
...
}
}
\source{
\url{https://data.giss.nasa.gov/gistemp/}
}
\usage{
GISTEMP
}
\description{
A dataset containing global temperature anomaly data.
}
\keyword{datasets}
