% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animateHTMLBy.R
\name{plotDensityByDimensionValue}
\alias{plotDensityByDimensionValue}
\title{Plot Density for a Chosen Dimension Value}
\usage{
plotDensityByDimensionValue(
  starsObj,
  attr,
  dim,
  value,
  title,
  sf = NULL,
  baseline = NULL,
  ylim = c(0, 0.5),
  xlim = NULL,
  sd = TRUE
)
}
\arguments{
\item{starsObj}{a stars object}

\item{attr}{attribute of interest}

\item{dim}{dimension of choice}

\item{value}{dimension value by which to plot}

\item{title}{title of plot}

\item{sf}{if not null, an sf object (e.g., a country from the rnaturalearth package) that will be plotted as an inset on the graph}

\item{baseline}{a vector of two dimension values that will provide a baseline density curve on the graph within the specified range.}

\item{ylim}{if specified, the range of y values for the density curve (by default, 0-0.5)}

\item{xlim}{if specified, the range of the x axis}

\item{sd}{how should the x axis range be determined programmatically? if TRUE (the default),
the x axis range is within three standard deviations of the attribute, if FALSE the range is the
minimum and maximum of the attribute.}
}
\value{

}
\description{
Plot Density for a Chosen Dimension Value
}
