% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_assign_colors.R
\name{st_assign_colors}
\alias{st_assign_colors}
\title{st_assign_colors}
\usage{
st_assign_colors(
  starsObj1,
  attrOf1,
  starsObj2,
  attrOf2,
  terciles,
  bounds = NULL,
  palette = NULL
)
}
\arguments{
\item{starsObj1}{a stars object}

\item{attrOf1}{which attribute of starsObj1 should be considered?}

\item{starsObj2}{a stars object}

\item{attrOf2}{which attribute of starsObj2 should be considered?}

\item{terciles}{should terciles be calculated?}

\item{bounds}{if custom bounds/levels are desired, default NULL}

\item{palette}{see Vizumap::build_palette if a custom palette is desired}
}
\value{
a data frame of the attributes, factor levels, and color (in hexadecimal)
}
\description{
Bin numeric data and assign colors for plotting
}
\examples{
out <- st_assign_colors(tmax, 1, tmin, 1, terciles = FALSE)
data <- out$data
colors <- out$colors
bounds <- out$bounds

}
