% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bivariate_contour_map.r
\name{st_bivariate_contour_map}
\alias{st_bivariate_contour_map}
\title{Produce points and density contour map for bivariate data}
\usage{
st_bivariate_contour_map(starsObj1, attr1, starsObj2, attr2)
}
\arguments{
\item{starsObj1}{a stars object}

\item{attr1}{attribute of starsObj1 to be plotted}

\item{starsObj2}{a stars object}

\item{attr2}{attribute of starsObj2 to be plotted}
}
\value{
a ggplot
}
\description{
Produce points and density contour map for bivariate data
}
\examples{
st_bivariate_contour_map(tmax[,,14613:18263],1,tmin[,,14613:18263],1)
}
