% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bivariate_histogram.R
\name{st_bivariate_histogram}
\alias{st_bivariate_histogram}
\title{st_bivariate_histogram}
\usage{
st_bivariate_histogram(data, ymax = NULL, title)
}
\arguments{
\item{data}{data frame containing data from two attributes, the binned
information and colour code.}

\item{ymax}{if custom maximum y-axis value is desired, default NULL}

\item{title}{title of plot}

\item{terciles}{should terciles be calculated?}
}
\value{
a ggplot
}
\description{
Plot bivariate spatio-temporal data
}
\details{
Assistance for the creation of bar chart levels was found at
https://stackoverflow.com/questions/10834382/ggplot2-keep-unused-levels-barplot
}
\examples{
# Preparing the data
out <- st_assign_colors(tmax, 1, tmin, 1, terciles = FALSE)
data <- out$data
colors <- out$colors
bounds <- out$bounds

# generating and plotting the histogram
bvh <- st_bivariate_histogram(data = data,
            title = "Max/Min temperature distribution")
bvh

# creating and attaching the key
key <- build_bkey(data, colors, terciles = TRUE)
view(key)
attach_key(bvh, key)

}
