% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bivariate_pointdensity.r
\name{st_bivariate_pointdensity}
\alias{st_bivariate_pointdensity}
\title{Produce point density plot of bivariate data}
\usage{
st_bivariate_pointdensity(data)
}
\arguments{
\item{data}{the data in long form, as returned from st_getAttributes and get_density.
The first two columns should be the variables of interest and the third should be their density}
}
\value{
a ggplot
}
\description{
Produce point density plot of bivariate data
}
