% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bivmap.R
\name{st_bivmap}
\alias{st_bivmap}
\title{st_bivmap}
\usage{
st_bivmap(data, ...)
}
\arguments{
\item{data}{a data frame as passed from st_assign_colors or assign_colors.df.
Columns for Longitude and Latitude are in the third and fourth columns, respectively, and
the color corresponding to each row lies in the sixth column.}

\item{...}{arguments passed to borders() from ggplot2. A convenient way to get borders from the maps
package onto the plot.}
}
\value{
a ggplot object
}
\description{
Plot bivariate spatial data. Grid points are colored according to the bivariate distribution.
First bin the data with st_assign_colors or assign_colors.df.
}
\examples{

}
