% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_change_key_bounds.R
\name{st_change_key_bounds}
\alias{st_change_key_bounds}
\title{st_change_key_bounds}
\usage{
st_change_key_bounds(key, bounds)
}
\arguments{
\item{key}{a key build using Vizumap::build_bkey}

\item{bounds}{the new bounds to be added}
}
\value{
a key, of class "bivkey"
}
\description{
Alter a key's bounds for display
}
\examples{
newBounds = c(-104, -2, 2, 202, -117, -2, 2, 150)
out = st_assign_colors(tmax, 1, tmin, 1, FALSE, bounds = newBounds)
data = out$data
colors = out$colors

bvh = st_bivariate_histogram(data = data, title = "Min/Max Temperature Distributions")
key = Vizumap::build_bkey(data, colors, terciles = FALSE)
key = st_change_key_bounds(key, newBounds)

attach_key(bvh, key)
}
