% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_assign_colors.R
\name{st_findNbounds}
\alias{st_findNbounds}
\title{Calculate bins for coloring}
\usage{
st_findNbounds(data, terciles)
}
\arguments{
\item{data}{dataset from which bounds are to be determined, column 1 is called "attr1" and column 2 is called "attr2"}

\item{terciles}{should terciles be calculated?}
}
\value{
vector of 8 with bounding values for the two variables
}
\description{
Calculate bins for coloring
}
