% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_getAttributes.r
\name{st_getAttributes}
\alias{st_getAttributes}
\title{st_getAttributes}
\usage{
st_getAttributes(starsObj1, attrOf1, starsObj2, attrOf2)
}
\arguments{
\item{starsObj1}{stars object}

\item{attrOf1}{attrOf1}

\item{starsObj2}{stars object}

\item{attrOf2}{attrOf2}
}
\description{
Function to extract the attributes of two stars objects
}
