% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmax.r
\docType{data}
\name{tmax}
\alias{tmax}
\title{Maximum Temperature}
\format{
A stars object where the spatial representation has 18263 rows and 1582 columns
and temporal dimension of length 7
}
\usage{
tmax
}
\description{
A stars object containing maximum temperature data. (A better description is needed here)
}
\keyword{datasets}
