% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animateHTMLBy.R
\name{animateHTMLBy}
\alias{animateHTMLBy}
\alias{anim_factory}
\title{Animate Densities by Dimension Values}
\usage{
animateHTMLBy(starsObj, attr, dim, title, anim_range = NULL, ...)

anim_factory(starsObj, attr, dim, title, anim_range = NULL, ...)
}
\arguments{
\item{starsObj}{stars object}

\item{attr}{attribute of interest}

\item{dim}{grouping dimension}

\item{title}{plot title}

\item{anim_range}{if specified, dimension values along which to animate}

\item{...}{arguments passed to plotDensityByDimensionValue()}
}
\value{

}
\description{
Animate Densities by Dimension Values
}
\examples{

}
