% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_key.R
\name{attach_key}
\alias{attach_key}
\title{View bivariate histograms with their keys}
\usage{
attach_key(plot, VizuKey, draw = TRUE)
}
\arguments{
\item{plot}{a plot such as those returned from st_bivariate_histogram()}

\item{VizuKey}{a key built with Vizumap::build_bkey()}

\item{draw}{should the plot be drawn on the current device (default)? if FALSE, returns a grob without drawing. See gridExtra::arrangeGrob()}
}
\value{
draws on the current device or returns a grob
}
\description{
View bivariate histograms with their keys
}
\examples{
plot = st_bivariate_histogram(tmax,1,tmin,1,terciles=FALSE,title="Max/Min temperature distribution")
palette = Vizumap::build_palette(name = "usr", colrange = list(colour = c("darkblue", "chartreuse4"), difC = c(3, 4)))
data = st_assign_colors(tmax,1,tmin,1,terciles=FALSE,palette=palette)
colnames(data) = c("Tmax","Tmin","bothVars","hex_code")
key = Vizumap::build_bkey(data, palette, terciles = TRUE)

attach_key(plot,key)
}
