% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgelineBy.R
\name{ridgelineBy}
\alias{ridgelineBy}
\title{Create Ridgeline Plots}
\usage{
ridgelineBy(starsObj, attr, dim, plotTitle, y_reverse = FALSE, sd = TRUE, ...)
}
\arguments{
\item{starsObj}{stars object}

\item{attr}{attribute of interest}

\item{dim}{grouping dimension}

\item{plotTitle}{title of plot}

\item{y_reverse}{should the y axis be reversed?}

\item{sd}{if TRUE (the default), the range of the x axis is within three standard deviations
of the mean, if FALSE the range of the x axis is the minimum and maximum of the attribute}

\item{...}{arguments passed to scale_y_continuous() or scale_y_reverse()}
}
\value{
a gg or ggplot object
}
\description{
Generate density plots on the same graph, grouped by a dimension of choice
}
\examples{
ridgelineBy(GISTEMP_Decadal, 1, 3, "Ridgeline", y_reverse = TRUE, breaks = seq(1970,2010,10))

# can subset stars objects by sf objects, such as countries from the rnaturalearth package
ridgelineBy(GISTEMP_Decadal[rnaturalearth::ne_countries(country = "Australia", returnclass = "sf")], 1, 3, "Australia Ridgeline", breaks = seq(1970,2010,10))
}
