% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bivariate_density.R
\name{st_2d_raster_density}
\alias{st_2d_raster_density}
\title{Produce density plots of bivariate data with 'raster' geom}
\usage{
st_2d_raster_density(data)
}
\arguments{
\item{data}{data}
}
\value{
a ggplot
}
\description{
Produce density plots of bivariate data with 'raster' geom
}
\examples{
# Preparing the data
data <- st_getAttributes(tmax[,,14613:18263], 1, tmin[,,14613:18263], 1)
rdens <- st_2d_raster_density(data)
rdens

}
