% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bivariate_density.R
\name{st_2d_raster_density}
\alias{st_2d_raster_density}
\title{Produce density plots of bivariate data with 'raster' geom}
\usage{
st_2d_raster_density(starsObj1, attr1, starsObj2, attr2)
}
\arguments{
\item{starsObj1}{a stars object}

\item{attr1}{attribute of starsObj1 to be plotted}

\item{starsObj2}{a stars object}

\item{attr2}{attribute of starsobj2 to be plotted}
}
\value{
a ggplot
}
\description{
Produce density plots of bivariate data with 'raster' geom
}
\examples{
st_2d_raster_density(tmax[,,14613:18263],1,tmin[,,14613:18263],1)
}
