% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bivariate_contour_map.r
\name{st_bivariate_contour_map}
\alias{st_bivariate_contour_map}
\title{Produce points and density contour map for bivariate data}
\usage{
st_bivariate_contour_map(data)
}
\arguments{
\item{data}{the data in long form, as returned from st_getAttributes or st_assign_colors.
The first two columns should be the variables of interest.}
}
\value{
a ggplot
}
\description{
Produce points and density contour map for bivariate data
}
\examples{
library(stars)
# matched time to decadal time periods
time <- as.numeric(st_get_dimension_values(tmax, 2))
periods <- list(c(1950:1959), c(1960:1969), c(1970:1979), c(1980:1989),c(1990:1999))
decades <- lapply(periods, function(x, time) which(time \%in\% x), time)

# We wish to view data from 1990-1999
data = st_getAttributes(tmax[,,decades[[5]]], 1, tmin[,,decades[[5]]], 1)
colnames(data) = c("MaxTemp", "MinTemp")
st_bivariate_contour_map(data)
}
