% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bivariate_histograms.R
\name{st_bivariate_histogram}
\alias{st_bivariate_histogram}
\title{Plot bivariate spatio-temporal data}
\usage{
st_bivariate_histogram(
  starsObj1,
  attr1,
  starsObj2,
  attr2,
  terciles,
  bounds = NULL,
  ymax = NULL,
  title
)
}
\arguments{
\item{starsObj1}{a stars object}

\item{attr1}{which attribute of starsObj1 should be plotted?}

\item{starsObj2}{a stars object}

\item{attr2}{which attribute of starsObj2 should be plotted?}

\item{terciles}{should terciles be calculated?}

\item{bounds}{if custom bounds should be plotted, default NULL}

\item{ymax}{if custom maximum y-axis value is desired, default NULL}

\item{title}{title of plot}
}
\value{
a ggplot
}
\description{
Plot bivariate spatio-temporal data
}
\examples{
st_bivariate_histogram(tmax,1,tmin,1,FALSE, title = "Max/Min temperature distribution")
}
