% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bivariate_density.R
\name{st_bivariate_pointdensity}
\alias{st_bivariate_pointdensity}
\title{Produce point density plot of bivariate data}
\usage{
st_bivariate_pointdensity(starsObj1, attr1, starsObj2, attr2)
}
\arguments{
\item{starsObj1}{a stars object}

\item{attr1}{attribute of starsObj1 to be plotted}

\item{starsObj2}{a stars object}

\item{attr2}{attribute of starsObj2 to be plotted}
}
\value{
a ggplot
}
\description{
Produce point density plot of bivariate data
}
\examples{
st_bivariate_pointdensity(tmax[,,14613:18263],1,tmin[,,14613:18263],1)
}
