% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bivariate_histograms.R
\name{st_multiple_histograms}
\alias{st_multiple_histograms}
\title{Create multiple plots of bivariate spatio-temporal data along chosen dimension values}
\usage{
st_multiple_histograms(
  starsObj1,
  attr1,
  starsObj2,
  attr2,
  dim_values,
  terciles,
  which = 1,
  titles,
  bounds = NULL
)
}
\arguments{
\item{starsObj1}{a stars object}

\item{attr1}{which attribute of starsObj1 should be plotted?}

\item{starsObj2}{a stars object}

\item{attr2}{which attribute of starsObj2 should be plotted?}

\item{dim_values}{the dimension values along which to create plots. Should be a list. Note that the actual indices of the dimension values are needed, see examples}

\item{terciles}{should terciles be calculated?}

\item{which}{which dimension is the one of interest? Should take a value between 1 and 3 (inclusive), use aperm() if need be}

\item{titles}{titles of the plots, should be of the same length as dim_values}

\item{bounds}{if custom bounds should be plotted, default NULL}
}
\value{

}
\description{
Create multiple plots of bivariate spatio-temporal data along chosen dimension values
}
\examples{
time = as.numeric(stars::st_get_dimension_values(tmax,2))
decades = list()
j=1
for(i in list(c(1950:1959),c(1960:1969),c(1970:1979),c(1980:1989),c(1990:1999))){
  decades[[j]] = which(time \%in\% i)
  j = j + 1
}
histograms = st_multiple_histograms(tmax,1,tmin,1,dim_values=decades,which=2,titles=c(1:5))
attach_key(gridExtra::arrangeGrob(histograms[[1]],histograms[[5]],ncol=1), key)
}
